requirements.txt
python-dotenv
httpx
python-telegram-bot==20.0
flask==2.1.2
gunicorn==20.1.0
